requirements_railway.txt
Flask==2.3.3
python-telegram-bot==20.3
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
